requirements.txt
Django==4.2
sqlparse==0.4.4
gunicorn==20.1.0
django-allauth==0.54.0
django-extensions==3.2.1
django-filter==23.1
django-js-asset==2.0.0
django-social-share==2.3.0
oauth2client==4.1.3
Pillow==9.4.0
python-dotenv==1.0.0
social-auth-app-django==5.2.0
social-auth-core==4.4.1